update route and refactor structure
