+++ src/App.js
@@ -4,11 +4,10 @@ import { Spin } from "antd";
 import "antd/dist/antd.css";
 import "./App.scss";
 import NotFound from "./components/not-found";
-import {Home, UserManagement, PostManagement, Test} from './features/dashboard'
 
 //Lazy loading
 const HomePage = React.lazy(() => import("./features/main"));
-const AdminPage = React.lazy(() => import("./features/dashboard"));
+const DashBoard = React.lazy(() => import("./features/dashboard"));
 
 function App() {
   
@@ -25,18 +24,8 @@ function App() {
           <Routes>
             <Route path="/" element={<Navigate to="/room-social-network" replace />} />
             <Route path="/room-social-network/*" element={<HomePage/>} />
+            <Route path="/admin-management/*" element={<DashBoard/>} />
             <Route path="*" element={<NotFound/>} />
-
-            <Route path="/room-social-network/admin" element={<AdminPage />}>           
-              <Route index element={<Navigate to="home" replace/>}/>
-              <Route path="home" element={<Home />} />
-              <Route path="quan-ly-nguoi-dung" element={<UserManagement />} />
-              <Route path="quan-ly-tin-dang" element={<PostManagement />} />
-              <Route path="giao-dien" element={<span>Tính năng đang được phát triển...</span>} />
-              <Route path="tu-khoa-hot" element={<span>Tính năng đang được phát triển...</span>} />
-           </Route>
-
-            <Route path="/room-social-network/admin/*" element={<NotFound />} />  
           </Routes>
         </BrowserRouter>
       </Suspense>
